modified API of package message and changed call id to Unix time in package rpc

+++ rpc/wrapper.go
@@ -2,13 +2,13 @@ package rpc
 
 // Call represents a remote call
 type call struct {
-	ID         uint32
+	ID         int64
 	CallerAddr string
 	Arg        interface{}
 }
 
 // Reply represents a reply to a remote call
 type reply struct {
-	ID  uint32
+	ID  int64
 	Ret interface{}
 }
